wip reconfigure handler test

+++ api/src/main/java/com/abiquo/api/services/cloud/VirtualMachineService.java
@@ -26,7 +26,6 @@ import static com.abiquo.api.util.URIResolver.buildPath;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.UUID;
@@ -34,7 +33,6 @@ import java.util.UUID;
 import javax.persistence.EntityManager;
 import javax.ws.rs.core.MultivaluedMap;
 
-import org.apache.commons.lang.StringUtils;
 import org.hibernate.Hibernate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -62,7 +60,6 @@ import com.abiquo.api.resources.cloud.VirtualDatacentersResource;
 import com.abiquo.api.services.DefaultApiService;
 import com.abiquo.api.services.NetworkService;
 import com.abiquo.api.services.RemoteServiceService;
-import com.abiquo.api.services.TaskService;
 import com.abiquo.api.services.UserService;
 import com.abiquo.api.services.VirtualMachineAllocatorService;
 import com.abiquo.api.services.stub.TarantinoJobCreator;
@@ -100,7 +97,6 @@ import com.abiquo.server.core.infrastructure.RemoteService;
 import com.abiquo.server.core.infrastructure.management.RasdDAO;
 import com.abiquo.server.core.infrastructure.management.RasdManagement;
 import com.abiquo.server.core.infrastructure.management.RasdManagementDAO;
-import com.abiquo.server.core.infrastructure.network.IpPoolManagement;
 import com.abiquo.server.core.infrastructure.network.IpPoolManagement;
 import com.abiquo.server.core.infrastructure.network.VLANNetwork;
 import com.abiquo.server.core.infrastructure.storage.DiskManagement;
@@ -167,9 +163,6 @@ public class VirtualMachineService extends DefaultApiService
     @Autowired
     private NetworkService ipService;
 
-    @Autowired
-    private TaskService tasksService;
-
     public VirtualMachineService()
     {
 
@@ -191,6 +184,7 @@ public class VirtualMachineService extends DefaultApiService
         this.appsLibRep = new AppsLibraryRep(em);
         this.tarantino = new TarantinoService(em);
         this.jobCreator = new TarantinoJobCreator(em);
+        this.ipService = new NetworkService(em);
     }
 
     public Collection<VirtualMachine> findByHypervisor(final Hypervisor hypervisor)
@@ -1421,7 +1415,7 @@ public class VirtualMachineService extends DefaultApiService
         tmp.setHighDisponibility(vm.getHighDisponibility());
         tmp.setHypervisor(vm.getHypervisor());
         tmp.setIdType(vm.getIdType());
-        tmp.setName(vm.getName());
+        tmp.setName("tmp_"+vm.getName());
         tmp.setPassword(vm.getPassword());
         tmp.setRam(vm.getRam());
         tmp.setState(VirtualMachineState.LOCKED);
@@ -1506,7 +1500,10 @@ public class VirtualMachineService extends DefaultApiService
         return tmp;
     }
 
-    /*
+    /**
+     * Gets the {@link VirtualMachine} backup created to store reconfigure previous state.
+     * 
+     * @return the virtualmachine with ''temp'' == provided vm identifier
      */
     public VirtualMachine getBackupVirtualMachine(final VirtualMachine vmachine)
     {
@@ -1521,34 +1518,25 @@ public class VirtualMachineService extends DefaultApiService
         return vmbackup;
     }
 
+    /**
+     * Cleanup backup resources
+     */
     public void deleteBackupResources(final VirtualMachine vm)
     {
         for (RasdManagement rasd : getBackupResources(vm))
         {
-            deleteResource(rasd);
-        }
-    }
-
-    protected List<RasdManagement> getBackupResources(final VirtualMachine rollbackVm)
-    {
-        try
-        {
-            rasdDao.enableTemporalOnlyFilter();
-
-            return rollbackVm.getRasdManagements();
-            // List<VolumeManagement> rollbackVolumes = rollbackVm.getv
-            // List<DiskManagement> rollbackDisks = rollbackVm.getDisks();
-            // List<IpPoolManagement> rollbackIps = rollbackVm.geti
-        }
-        finally
-        {
-            rasdDao.disabledTemporalOnlyFilter();
+            removeResource(rasd, true);
         }
     }
 
     /**
-    * 
-    * */
+     * Updates all the attributes and resource attachments of ''updatedVm'' from the backup
+     * ''rollbackVm''.
+     * 
+     * @param updatedVm, current state of the virtual machine (not applied in the hypervisor)
+     * @param rollbackVm, state of updaedVm previous to the reconfigure.
+     * @return updatedVm with the attributes and resource attachments of rollbackVm.
+     */
     public VirtualMachine restoreBackupVirtualMachine(final VirtualMachine updatedVm,
         final VirtualMachine rollbackVm)
     {
@@ -1579,42 +1567,51 @@ public class VirtualMachineService extends DefaultApiService
         List<RasdManagement> updatedResources = updatedVm.getRasdManagements();
         List<RasdManagement> rollbackResources = getBackupResources(rollbackVm);
 
-        // for updated if not temp -> delete rasd (if network) + delte / else update form temp
-        // (settemp=null)
         for (RasdManagement updatedRasd : updatedResources)
         {
             RasdManagement rollbackRasd = getBackupResource(rollbackResources, updatedRasd.getId());
-            deleteResource(updatedRasd);
 
-            if (rollbackRasd != null) // restore the original resource attachment
+            if (rollbackRasd == null)
             {
-                restoreRollbackResouce(rollbackRasd, updatedVm);
+                LOGGER.trace("restore: detach resource "+ updatedRasd.getId());
+                removeResource(updatedRasd, false);
             }
-            // else the original vm do not include this resource (updatedRasd already deleted)
-
         }
 
-        // for rollback if temp == 0 --> temp == null + attach a new machine
         for (RasdManagement rollbackRasd : rollbackResources)
         {
-            // updated vm do not include this resouces (deatach), we need to restore it
-            if (rollbackRasd.getTemporal() == 0)
+            RasdManagement originalRasd =
+                getOriginalResource(updatedResources, rollbackRasd.getTemporal());
+
+            if (!originalRasd.isAttached() || originalRasd.getVirtualMachine() == null) // XXX
             {
-                restoreRollbackResouce(rollbackRasd, updatedVm);
+                LOGGER.trace("restore: attach resource "+ originalRasd.getId());
+                restoreResource(originalRasd, updatedVm, rollbackRasd.getRasd().getGeneration().intValue());
             }
+
+            removeResource(rollbackRasd, true);
         }
 
         rasdDao.flush(); // update virtual machine resources
-
-        repo.update(updatedVm); // remove orphans !!!
+        repo.update(updatedVm);
         repo.deleteVirtualMachine(rollbackVm);
 
         // TODO check temporal not set in updatedVm
 
+        LOGGER.info("restored virtual machine {} from backup", updatedVm.getUuid());
+
         return updatedVm;
     }
 
-    private void deleteResource(final RasdManagement rasd)
+    /** Re attach the resource to the virtual machine */
+    private void restoreResource(final RasdManagement rasd, final VirtualMachine updatedVm, final Integer originalSequence)
+    {
+        rasd.attach(originalSequence, updatedVm);
+        rasd.setVirtualMachine(updatedVm); // XXX
+    }
+    
+    /** Removes the resource attachment */
+    private void removeResource(final RasdManagement rasd, final boolean remove)
     {
         rasd.detach();
         rasd.setVirtualMachine(null); // FIXME until detach is not implemented in all
@@ -1625,26 +1622,57 @@ public class VirtualMachineService extends DefaultApiService
             rasdRawRao.remove(rasd.getRasd());
         }
 
-        rasdDao.remove(rasd);
+        if(remove)
+        {            
+            rasdDao.remove(rasd);
+        }
     }
 
-    private void restoreRollbackResouce(final RasdManagement rasd, final VirtualMachine vm)
+    /**
+     * Get the resources attached to the provided backup virtualmachine.
+     */
+    private List<RasdManagement> getBackupResources(final VirtualMachine rollbackVm)
     {
-
-        // using the original sequence
-        final Integer originalSequence = rasd.getRasd().getGeneration().intValue();
-        rasd.attach(originalSequence, vm);
-
-        rasd.setTemporal(null);
+        try
+        {
+            rasdDao.enableTemporalOnlyFilter();
+            return rollbackVm.getRasdManagements();
+        }
+        finally
+        {
+            rasdDao.disabledTemporalOnlyFilter();
+        }
     }
 
-    /** Find */
+    /**
+     * Find the backup resources with temporal pointing to the provided resource identifier.
+     * 
+     * @return resource with temporal == provided resource id, null if not found
+     */
     private RasdManagement getBackupResource(final List<RasdManagement> rollbackResources,
         final Integer tempRasdManId)
     {
         for (RasdManagement rasdman : rollbackResources)
         {
             if (tempRasdManId.equals(rasdman.getTemporal()))
+            {
+                return rasdman;
+            }
+        }
+        return null;
+    }
+
+    /**
+     * Find the original resource with identifier equals to the provided backup resource temporal.
+     * 
+     * @return resource with id == backup.temporal, null if not found
+     */
+    private RasdManagement getOriginalResource(final List<RasdManagement> updatedResources,
+        final Integer temporalId)
+    {
+        for (RasdManagement rasdman : updatedResources)
+        {
+            if (temporalId.equals(rasdman.getId()))
             {
                 return rasdman;
             }
